+++ __init__.py
@@ -19,6 +19,7 @@ time = shared.time
 key_ = shared.key_
 mouse_ = shared.mouse_
 
+print('Current Expy: ', os.path.dirname(os.path.abspath(__file__)))
 
 def start(setting_file='setting.txt', fullscreen=True, winsize=(800, 600), mouse_visible=False, normal_font_size=20, stim_font_size=None, distance=60, diag=23, angel=2.5, font_color=C_white, background_color=C_gray, sample_rate=44100, port='', vsync=True):
     '''
@@ -61,7 +62,6 @@ def start(setting_file='setting.txt', fullscreen=True, winsize=(800, 600), mouse
     -------
     None
     '''
-
     'Parameters'
     func_var = locals().copy()
     shared.setting = readSetting(setting_file)